refactor(web): split test file to match module structure
Move tests from the monolithic types.test.ts into routing.test.ts and
projects.test.ts, matching the new module boundaries. Extract shared
makeSession helper into test-helpers.ts. Delete types.test.ts (types.ts
is now pure interfaces with nothing to test; normalizeRemote tests moved
to projects.test.ts).

All 236 tests pass across 11 test files.
