+++ apps/gmux-web/src/test-helpers.ts
@@ -1,0 +1,20 @@
+import type { Session } from './types'
+
+export function makeSession(overrides: Partial<Session> & { id: string; cwd: string }): Session {
+  return {
+    created_at: new Date().toISOString(),
+    command: ['pi'],
+    kind: 'pi',
+    alive: true,
+    pid: 1234,
+    exit_code: null,
+    started_at: new Date().toISOString(),
+    exited_at: null,
+    title: 'test',
+    subtitle: '',
+    status: null,
+    unread: false,
+    socket_path: '/tmp/test.sock',
+    ...overrides,
+  }
+}
